+++ pkg/util/exprfilter/exprperf_test.go
@@ -1,4 +1,9 @@
-package run
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+package exprfilter
 
 import (
 	"testing"
